# Added the ability to move NPCs via script
Added the Lua function “walkNumTilesWithDirection”, this allows a user
to specify the npc name, number of tiles and direction to move. Note
that it currently only supports moving in a single direction at a time,
so if you want to make a complex movement you will have to chain calls
together.