# Added NPC auto walk logic
NPCs can now be given the ability to move around on their own. This can
either be done by defining a rectangle for their walk bounds in script,
or by defining the box in the map editor. The NPC also needs to know
how long to wait between steps, hence the function setStepDelay in the
NPC lua class, this delay can also be set like other properties in the
map editor.