# Re-did player movement input
I’ve made player movement to be like NPC movement, where the player
will always move 1 tile in whatever direction they last pressed the
arrow. Previously you would only move an amount equal to the delta time
* movement speed while the movement was held down. This is probably
better for a more SNES type movement scheme, but I think for simplicity
and to capture the NES type system we are going for this makes more
sense. I left the other code in for now, although it might actually
take more work than just uncommenting it to make it work again. It
could be worth it to have a flag to switch between the two I suppose.