# Created Yes/No dialogue boxes
Can be called from LUA by using the function talkYesNo and passing in
the text strings for Yes and No responses. I also made it so the hud
layer is hidden while dialogue is displayed. And I made it so NPCs
won’t move around when you are talking to them.