# Finally got Lua bindings working
Finally got the stupid Lua bindings stuff to work by installing the
android NDK. I’ve generated bindings for the NPCSprite and ActionSprite
classes. This supersedes the previous scripting interface for NPCs
(this works better I promise). I’m in the process of redoing the sample
cut scene in the throne room to show off this reworked interface.